Workarounds to allow dagger's usage in SDK 24
Started with requiring 1.8 to compile against SDK 24.
Then Jack toolchain is required for 1.8
Jack doesn't support annotations, so dagger blows up.
Fix is to update to 2.2.0-alpha5 of gradle tools that supports
annotation processing
Then dagger and dagger-compiler had to be updated to version 2.5

But, now android-apt is useless, so 1 less dependancy