Behaviour subject works as expected with tests
The idea is you can subscribe to the observable that tells you
what state the refresh is in at any time. If you leave the screen
and come back, it doesn't matter, you're completely stateless.

The other part is if you want to refresh, you just push the PublishSubject
and you never have to worry about it again, the observable mentioned above
gets updated correctly

Also, tests works. Only caveat is the tests have a `wait` style problem
because Rx takes time. TestObserver helps but its not meant for infinite streams
which is what we're using