CalendarModel Unit tests are now deterministic
A new Android Scheduler Rule is now available but
ultimately wasn't needed.

Fixed it by creating a new blocking observable that collects
all the items into a list after x amount of time.
This ensures, ONLY the observables we expect are sent.

We block, until its returned and assert the results