Fixed an issue where Booking interaction wasn't loading on app coma state
The reason is that the drawer is handled fine, and before, the assumption
was that the drawer would govern everything, so it was handled. However
now there's non-drawer screens which handle their own transaction state and
backstack. In this case this is a bit of a hack. Hopefully if it continues
there migth be a need to refactor this all so theres only 1 root to the app.
Instead of using the drawer as the root